Rename, add typescript, move to avoid dep cycle
